Fix: Lambda 파이프라인 로드 문제 해결
